refactor ij by pkfk
